Properly handle NO DATA with  multiple conditions with a mix of aggregations and document count thresholds (#154690)
This PR fixes #154553, which arises when a rule containing
aggregation-based conditions is combined with document count conditions.
In such a case, the aggregation-based conditions may return "NO DATA".
Unlike document count-based conditions, which do not typically generate
a "NO DATA" response since this is equivalent to zero and only triggers
if the user defines the condition as `Document count < 0`;
aggregation-based conditions produce a `null` result that is interpreted
as "NO DATA".

This PR also improves the "NO DATA" reason message when a group-by field
is not defined. The reason message before the change looked like
`test.metric.3 reported no data in the last 1m for ` which I changed to
`test.metric.3 reported no data in the last 1m` to be more readable.

This also fixes the context variables for document count-based
conditions by setting the `metric` context variable to `count`.